fix Windows Visual Studio warnings

+++ src/ssl.c
@@ -24055,7 +24055,7 @@ void* wolfSSL_GetRsaDecCtx(WOLFSSL* ssl)
         (void)depth;
         WOLFSSL_STUB("wolfSSL_set_verify_depth");
 #else
-        ssl->options.verifyDepth = depth;
+        ssl->options.verifyDepth = (byte)depth;
 #endif
     }
